Commit to 7 files
FINALLY got the tilemap engine to work, though it is not fully
optimised. Takes about 20 seconds to load, so we might need a loading
screen. Though, we got the tilemap engine to work!!!!
Threads will do, but that will be for another time.
Also, changed the tilemap, so that each tile is bigger. The graphics are
still crude, though (like always).